add validation address model
